+++ TabletopTweaks/Bugfixes/Items/Weapons.cs
@@ -19,6 +19,7 @@ using System.Reflection;
 using System.Reflection.Emit;
 using TabletopTweaks.Config;
 using TabletopTweaks.Extensions;
+using TabletopTweaks.Utilities;
 
 namespace TabletopTweaks.Bugfixes.Items {
     static class Weapons {
@@ -31,8 +32,11 @@ namespace TabletopTweaks.Bugfixes.Items {
                 Initialized = true;
 
                 Main.LogHeader("Patching Weapons");
+
                 PatchBladeOfTheMerciful();
                 PatchHonorableJudgement();
+                PatchTerribleTremble();
+
                 PatchThunderingBurst();
 
                 void PatchBladeOfTheMerciful() {
@@ -94,6 +98,26 @@ namespace TabletopTweaks.Bugfixes.Items {
                         };
                     });
                     Main.LogPatch("Patched", JudgementOfRuleEnchantment);
+                }
+                void PatchTerribleTremble() {
+                    if (ModSettings.Fixes.Items.Weapons.IsDisabled("TerrifyingTremble")) { return; }
+
+                    var TerrifyingTrembleItem = Resources.GetBlueprint<BlueprintItemWeapon>("8c31891423c4405393741e829aebec85");
+                    var Enhancement5 = Resources.GetBlueprint<BlueprintWeaponEnchantment>("bdba267e951851449af552aa9f9e3992");
+                    var Ultrasound = Resources.GetBlueprint<BlueprintWeaponEnchantment>("582849db96824254ebcc68f0b7484e51");
+                    var TerrifyingTrembleEnchant_TTT = Resources.GetModBlueprint<BlueprintWeaponEnchantment>("TerrifyingTrembleEnchant_TTT");
+
+                    TerrifyingTrembleItem.m_DescriptionText = Helpers.CreateTaggedString($"{TerrifyingTrembleItem.name}.description", 
+                        "Whenever the wielder of this +5 ultrasound earthbreaker lands a killing blow, he deals sonic damage equal to his ranks in " +
+                        "the Athletics skill to all enemies within 10 feet. Successful Reflex save (DC 30) halves the damage.");
+
+                    TerrifyingTrembleItem.m_Enchantments = new BlueprintWeaponEnchantmentReference[] {
+                        Enhancement5.ToReference<BlueprintWeaponEnchantmentReference>(),
+                        Ultrasound.ToReference<BlueprintWeaponEnchantmentReference>(),
+                        TerrifyingTrembleEnchant_TTT.ToReference<BlueprintWeaponEnchantmentReference>(),
+                    };
+
+                    Main.LogPatch("Patched", TerrifyingTrembleItem);
                 }
                 void PatchThunderingBurst() {
                     if (ModSettings.Fixes.Items.Weapons.IsDisabled("ThunderingBurst")) { return; }